backend/requirements.txt
fastapi==0.109.0
uvicorn==0.24.0
groq==0.4.2
python-dotenv==1.0.0
chromadb==0.4.22
sentence-transformers==2.2.2
python-multipart==0.0.6
aiofiles==23.2.1
sqlalchemy==2.0.23
